Spy of User Repository for get users function was created as a promise of empty array, which is going to be resolved on calling this.
Through this way, the behavior of the method can be tested by recreating the condition where the list is empty in database. In spying, the original method replaced with a mock function and change its implementation to return the value as per our need.
After running the test the spy function to the original restored by mockRestore function.